# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ISM Tecnologia Financial System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/projecao-dias:
    get:
      operationId: getDashboardProjecaoDias
      tags: [dashboard]
      parameters:
        - name: dias
          in: query
          schema:
            type: integer
            enum: [15, 30, 60]
            default: 30
      summary: Cash flow projection for next N days
      responses:
        "200":
          description: Projection data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjecaoDiaItem"
  /dashboard/inadimplencia-clientes:
    get:
      operationId: getDashboardInadimplenciaClientes
      tags: [dashboard]
      parameters:
        - name: tab
          in: query
          schema:
            type: string
            enum: [inadimplente, proximos_vencer, vencidos]
            default: inadimplente
      responses:
        "200":
          description: Client overdue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InadimplenciaItem"
  /dashboard/inadimplencia-fornecedores:
    get:
      operationId: getDashboardInadimplenciaFornecedores
      tags: [dashboard]
      parameters:
        - name: tab
          in: query
          schema:
            type: string
            enum: [inadimplente, proximos_vencer, vencidos]
            default: inadimplente
      responses:
        "200":
          description: Supplier overdue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InadimplenciaItem"
...
  /dashboard/fluxo-caixa-mensal:
    get:
      operationId: getDashboardFluxoCaixaMensal
      tags: [dashboard]
      parameters:
        - name: ano
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly cash flow bar chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FluxoCaixaMensalItem"
...
  /filiais:
    get:
      operationId: getFiliais
      tags: [filiais]
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Filial"
    post:
      operationId: createFilial
      tags: [filiais]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFilial"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filial"
  /filiais/{id}:
    put:
      operationId: updateFilial
      tags: [filiais]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFilial"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filial"
    delete:
      operationId: deleteFilial
      tags: [filiais]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /departamentos:
    get:
      operationId: getDepartamentos
      tags: [departamentos]
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Departamento"
    post:
      operationId: createDepartamento
      tags: [departamentos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartamento"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Departamento"
  /departamentos/{id}:
    put:
      operationId: updateDepartamento
      tags: [departamentos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartamento"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Departamento"
    delete:
      operationId: deleteDepartamento
      tags: [departamentos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /centros-custos:
    get:
      operationId: getCentrosCustos
      tags: [departamentos]
      responses:
        "200":
          description: List of cost centers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CentroCusto"
    post:
      operationId: createCentroCusto
      tags: [departamentos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCentroCusto"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CentroCusto"
  /centros-custos/{id}:
    put:
      operationId: updateCentroCusto
      tags: [departamentos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCentroCusto"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CentroCusto"
    delete:
      operationId: deleteCentroCusto
      tags: [departamentos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /v1/filiais:
    get:
      operationId: apiV1Filiais
      tags: [tokens-api]
      summary: PowerBI - List branches
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Branches list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Filial"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ProjecaoDiaItem:
      type: object
      properties:
        data:
          type: string
        saldo:
          type: number
        receber:
          type: number
        pagar:
          type: number
      required:
        - data
        - saldo
        - receber
        - pagar
    InadimplenciaItem:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        valor:
          type: number
        vencimento:
          type: string
      required:
        - id
        - nome
        - valor
        - vencimento
...
    FluxoCaixaMensalItem:
      type: object
      properties:
        mes:
          type: string
        entradas:
          type: number
        saidas:
          type: number
      required:
        - mes
        - entradas
        - saidas
...
    Filial:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
      required:
        - id
        - nome
    CreateFilial:
      type: object
      properties:
        nome:
          type: string
      required:
        - nome
    UpdateFilial:
      type: object
      properties:
        nome:
          type: string
...
    Departamento:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
      required:
        - id
        - nome
    CreateDepartamento:
      type: object
      properties:
        nome:
          type: string
      required:
        - nome
    UpdateDepartamento:
      type: object
      properties:
        nome:
          type: string
    CentroCusto:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        departamento_id:
          type: integer
      required:
        - id
        - nome
    CreateCentroCusto:
      type: object
      properties:
        nome:
          type: string
        departamento_id:
          type: integer
      required:
        - nome
    UpdateCentroCusto:
      type: object
      properties:
        nome:
          type: string
        departamento_id:
          type: integer